+++ includes/class-plugin.php
@@ -46,6 +46,7 @@ class Plugin {
 		add_action( 'init', array( $this, 'load_textdomain' ) );
 
 		( new \FsnwKeyManagement\Frontend\FrontendController() )->init();
+		( new Integrations\SignatureKioskIntegration() )->init();
 	}
 
 	/**